[cli] Splits profile.rs into interface and cliclack implementation files
