+++ crates/goose-cli/src/prompt/mod.rs
@@ -1,0 +1,2 @@
+pub mod cliclack;
+pub mod prompt;
